make the tests run in parallel
I ran "make check" in libqb and watched impatiently ;-)
as the tests ran in serial on my multi-core system.
If you add automake's "parallel-tests" option below, they'll
run in parallel.

The color-tests option makes it so the "PASS" and "FAIL" words
are colored green and red respectively.  Nice, but no big deal.
Similarly nice-to-have is the "dist-xz" option.
That makes is so when you run "make dist" it creates xz-compressed
tarballs in addition to the usual gzip-compressed ones.
The advantage of also using xz is that it compresses significantly better:

    $ du -sh *z
    1.2M    libqb-0.4.1.75-3737-dirty.tar.gz
    812K    libqb-0.4.1.75-3737-dirty.tar.xz

Now, whenever a project releases both gzip-compressed and
xz-compressed tarballs, I always download only the latter, smaller ones.
Not only does it save time on the download (minimal in this case),
but it occupies less space on disk and uncompresses faster.

Signed-off-by: Angus Salkeld <asalkeld@redhat.com>