avoid autoconf warning
Autoconf warned about this:
configure.ac:72: warning: AC_LANG_CONFTEST: no AC_LANG_SOURCE\
  call detected in body
That's warning about the use via AC_COMPILE_IFELSE without
specifying a language.  The easiest way to work around that is
to avoid the use of CC altogether and instead to use the preprocessor.
* configure.ac (cc_supports_flag): Use AC_PREPROC_IFELSE in place
of AC_COMPILE_IFELSE (and CPPFLAGS in place of CFLAGS).

Signed-off-by: Angus Salkeld <asalkeld@redhat.com>