LOG: copy the function/filename for dynamic callsites
callsites originate from the assumption that the function/filename/format
are all statically allocated. When moving to dynamic callsite we kept
this assumption. But people are now passing in function/file names that
later get free'd. So we need to make sure that they stay persistant by
allocating them ourselves.

Signed-off-by: Angus Salkeld <asalkeld@redhat.com>