ipc: Don't send the dispatch_del() function a closed fd
The current code closes the fd before passing it to the
provided delete function. In the default case this can generate
"epoll_ctl(del): Bad file descriptor" messages when debug logging
is enabled.

This patch swaps over the calls so that a valid FD is passed to
the delete function.

Signed-off-by: Christine Caulfield <ccaulfie@redhat.com>
Reviewed-By: Ken Gaillot <kgaillot@redhat.com>