Fix address.sun_path in lib/ipc_setup.c and lib/ipc_socket.c
This patch fixes writing of the *.sun_path socket names to become
non-zero for bind() and connect().

The previous race condition is back for the ipc_stress_connections_us
test. Adding the unlink statement solves this problem (unknown how,
since res=-1 due to ENOENT: No such file or directory.)

This patch and problem was reported to Debian in bug #803776, see
http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=803776

* lib/ipc_setup.c: Fix correct printing of *.sun_path
* lib/ipc_socket.c: Likewise