build: release.mk: simplify/generalize GPG signing rule
Pattern matching is a brilliant DRY technique in makefile's composition,
so it would be a pity to miss such a nonconflicting application related
to the previous, "sign tarballs" commit.

For that to work reliably for the complete dependency backtrack, though,
a pre-existing issue with parallelized work on the prerequisites needs
to be fixed first.  The problem was "tarballs" (phony) target missing
in the linearized supply chain for the signed tarballs, as opposed to
to supply chain for the checksums file that can be worked on in an
(unordered!) interleave with the work on these signed tarballs.
Hence fix this race condition by stretching "tarballs" target as
progress barrier (using it as a prerequisite) uniformly for all the
deliverables that are going to be signed (used to be just for "sha256"
target)).  And when at it, make "deliverables" variable list them
conveniently (along with reformatting definitions of other variables
for uniformity).

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>