build: release.mk: fix no-release conflict (implied-required version)
Previously, generating custom versioned tarballs without requiring
a "release" (which comprises also tagging hence in turn making the
implied version match the required one as desired) would not allow
to test the release routines thoroughly incl. GPG signing.
Unfortunately, allowing for a temporary tag is prone to accidentally
slipping it to the distributed tree, so that should rather be avoided.
Now it can.

Another step to blur the boundaries between whether "release" is
specified or not is make checksums file generated unconditionally.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>