build: configure: check section boundary symbols present in the test
There was an idea to make apparently run-time test set to that effect,
but it would make cross-building harder.  So arrange the test as if it
would be meant for AC_TRY_RUN, but achieve the same as with the first
assertion by the means of inspecting the linked result with, possibly
target-specific, nm utility.

While arranging the test for AC_TRY_RUN, based on feedback by Ferenc
Wágner, unify and increase usability of the run-time error signalling
through assertions.

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>