Monotonic clock support requires runtime check
_POSIX_MONOTONIC_CLOCK is defined as 0 on Linux, calling for a runtime
check.  After that we fall back to CLOCK_REALTIME if needed.  In this
commit for clock_getres only, which failed on Hurd glibc until 2.24-18
because of an omission: other functions already had an internal fallback
to CLOCK_REALTIME.  CLOCK_REALTIME is guaranteed to work by POSIX.1-2001;
CLOCK_MONOTONIC is only guaranteed to be a known clockid_t.

Signed-off-by: Ferenc Wágner <wferi@debian.org>